==> 1_getMacData.ipynb <==
import os
import pandas as pd
from datetime import datetime, timedelta, timezone
from pymongo import MongoClient
from bson import ObjectId
from reportlab.lib import colors
from reportlab.lib.pagesizes import letter
from reportlab.platypus import SimpleDocTemplate, Paragraph, Spacer, Table, TableStyle
from reportlab.lib.styles import getSampleStyleSheet
from reportlab.lib.styles import getSampleStyleSheet
import pytz
import json
import shutil

styles = getSampleStyleSheet()

# MongoDB configuration
MONGO_DB_NAME = "admin"
MONGO_URL = f"mongodb://wiser:wiser123@115.186.183.129:50001,100.28.143.179:27018,100.28.143.179:27019/admin?replicaSet=wiserReplica"

mac_address = "08:D1:F9:A7:81:B8"

# Connect to MongoDB


def connect_mongo():
    client = MongoClient(MONGO_URL)
    return client[MONGO_DB_NAME]

# Fetch lab name from MongoDB


# Fetch data for a specified date range from MongoDB

def fetch_data_for_date_range(collection, query):
    entries_data = collection.find(query)
    entries_df = pd.DataFrame(list(entries_data))
    return entries_df

def fetch_utilization_data(mac):
    # Connect to MongoDB
    db = connect_mongo()

    # ------------------------------ Current Data ------------------------------
    current_cursor = db['cts'].find({"mac": mac}).sort([("created_at", 1)])
    current_data = list(current_cursor)  # Convert cursor to list

    if current_data:  # Check if the list is not empty
        current_data = pd.DataFrame(current_data)  # Convert to DataFrame

    # ------------------------------ Vibration Data ------------------------------
    vibration_cursor = db['vibrations'].find(
        {"mac": mac}).sort([("created_at", 1)])
    vibration_data = list(vibration_cursor)

    if vibration_data:
        vibration_data = pd.DataFrame(vibration_data)

    # ------------------------------ Temperature Data ------------------------------
    temperature_cursor = db['thermisters'].find(
        {"mac": mac}).sort([("created_at", 1)])
    temperature_data = list(temperature_cursor)

    if temperature_data:
        temperature_data = pd.DataFrame(temperature_data)

    # Create 'RawData' directory if it doesn't exist
    if os.path.exists("RawData"):
        shutil.rmtree("RawData")  # Remove the old data
    os.makedirs("RawData")

    # Save data to CSV
    if not current_data.empty:
        current_data.to_csv("RawData/current_data.csv", index=False)
    if not vibration_data.empty:
        vibration_data.to_csv("RawData/vibration_data.csv", index=False)
    if not temperature_data.empty:
        temperature_data.to_csv("RawData/temperature_data.csv", index=False)

    return True

# Example usage
fetch_utilization_data(mac_address)

# make a json file named 'model_data.json' and write the mac address in it

data = {}

data['mac'] = mac_address

# if file exists, delete it
if os.path.exists('model_data.json'):
    os.remove('model_data.json')

with open('model_data.json', 'w') as json_file:
    json.dump(data, json_file)
