==> 2_Pre_Data_Processing.ipynb <==
#import libraries
import glob
from keras.models import Sequential, load_model
import numpy as np
import pandas as pd
from keras.layers import Dense
from sklearn.model_selection import train_test_split
from sklearn.preprocessing import LabelEncoder, MinMaxScaler
import matplotlib.pyplot as plt
import keras as k
import datetime
from datetime import datetime, timedelta

# Importing the dataset
dataset = pd.read_csv('./RawData/current_data.csv')

dataset.head()


dataset.shape

#Creating a list of column names to keep
columns_to_retain = ['CT1', 'CT2', 'CT3', 'CT_Avg', 'total_current', 'state', 'created_at']

#Drop the columns that are not wanted in columns_to_retain
dataset = dataset[columns_to_retain]

pd.DataFrame(dataset)

# Drop rows where state is not ON
dataset = dataset[dataset.state != 'OFF']

# drop idle
dataset = dataset[dataset.state != 'IDLE']

pd.DataFrame(dataset)

# Convert created_at 2024-03-26 04:17:37 format to 2024-08-11T20:29:06.000Z format

# Convert to datetime object ie datetime.strptime


dataset['created_at'] = pd.to_datetime(dataset['created_at'])

pd.to_timedelta(dataset.groupby('created_at').cumcount(), unit='ms')

dataset['created_at'] = dataset['created_at'] + pd.to_timedelta(dataset.groupby('created_at').cumcount(), unit='ms')


dataset['created_at'] = dataset['created_at'].dt.strftime('%Y-%m-%dT%H:%M:%S.%fZ')




pd.DataFrame(dataset)

# remove cleandata folder and make new one

import os

os.system('rm -rf CleanData')
os.system('mkdir CleanData')




dataset['created_at'].value_counts()

# Saving the dataset to a new csv file

dataset.to_csv('./CleanData/current_data.csv', index=False)

# Load the new dataset
dataset = pd.read_csv('./RawData/vibration_data.csv')

dataset.head()

#Creating a list of column names to keep
columns_to_retain = ['vibration', 'created_at']

#Drop the columns that are not wanted in columns_to_retain
dataset = dataset[columns_to_retain]

pd.DataFrame(dataset)

# Drop rows where vibration is zero or missing
dataset = dataset[dataset.vibration != 0]

# dropna function is used to remove missing values
dataset = dataset.dropna(axis=0)

pd.DataFrame(dataset)

# Convert created_at 2024-03-26 04:17:37 format to 2024-08-11T20:29:06.000Z format

# Convert to datetime object ie datetime.strptime


dataset['created_at'] = pd.to_datetime(dataset['created_at'])

pd.to_timedelta(dataset.groupby('created_at').cumcount(), unit='ms')

dataset['created_at'] = dataset['created_at'] + pd.to_timedelta(dataset.groupby('created_at').cumcount(), unit='ms')


dataset['created_at'] = dataset['created_at'].dt.strftime('%Y-%m-%dT%H:%M:%S.%fZ')



dataset['created_at'].value_counts()

# Saving the dataset to a new csv file
dataset.to_csv('./CleanData/vibration_data.csv', index=False)

# loading data from the csv files
dataset = pd.read_csv('./RawData/temperature_data.csv')

dataset.head()

#Creating a list of column names to keep
columns_to_retain = ['therm_temp', 'created_at']

#Drop the columns that are not wanted in columns_to_retain
dataset = dataset[columns_to_retain]

pd.DataFrame(dataset)


# Drop rows where vibration is zero or missing
dataset = dataset[dataset.therm_temp != 0]

# dropna function is used to remove missing values
dataset = dataset.dropna(axis=0)

pd.DataFrame(dataset)

# Convert created_at 2024-03-26 04:17:37 format to 2024-08-11T20:29:06.000Z format

# Convert to datetime object ie datetime.strptime


dataset['created_at'] = pd.to_datetime(dataset['created_at'])

pd.to_timedelta(dataset.groupby('created_at').cumcount(), unit='ms')

dataset['created_at'] = dataset['created_at'] + pd.to_timedelta(dataset.groupby('created_at').cumcount(), unit='ms')


dataset['created_at'] = dataset['created_at'].dt.strftime('%Y-%m-%dT%H:%M:%S.%fZ')



dataset['created_at'].value_counts()

# Saving the dataset to a new csv file
dataset.to_csv('./CleanData/temperature_data.csv', index=False)

# loading cleaned datasets and merging them together interms of time
cts = pd.read_csv('./CleanData/current_data.csv')
temps = pd.read_csv('./CleanData/temperature_data.csv')
vibrations = pd.read_csv('./CleanData/vibration_data.csv')

# Merge the datasets
data = pd.merge(cts, temps, on='created_at')

data = pd.merge(data, vibrations, on='created_at')

pd.DataFrame(data)

# Dropping the rows with state as 'IDLE'
data = data[data.state != 'IDLE']

pd.DataFrame(data)

# Dropona function is used to remove missing values

data = data.dropna(axis=0)

pd.DataFrame(data)

data['created_at'].value_counts()

# Saving the dataset to a new csv file
data.to_csv('./CleanData/data.merged.csv', index=False)
