==> 5_evaluate_model.ipynb <==
#import libraries
import glob
from keras.models import Sequential, load_model
import numpy as np
import pandas as pd
from keras.layers import Dense
from sklearn.model_selection import train_test_split
from sklearn.preprocessing import LabelEncoder, MinMaxScaler
import matplotlib.pyplot as plt
import keras as k
import datetime
import joblib

df = pd.read_csv('./DataSet/data.merged.csv')

df.head()

pd.DataFrame(df)

#Split the data into independent (X) data set -features  and (Y) data set -targets
X = df.drop(['classification'], axis=1)
# X =X.drop(['index'], axis=1)
Y = df['classification']

# X = X.drop(['created_at'], axis=1)

pd.DataFrame(X)

#Feature Scaling
#min-max scaler method scales the data set so that all the input features lie between 0 and 1
X_scaler = MinMaxScaler()
X_scaler.fit(X)
column_names = X.columns
X[column_names] = X_scaler.transform(X)

# saving scaler
joblib.dump(X_scaler, './Models/scaler.pkl')

pd.DataFrame(X)

# load model from file
model = load_model('./models/model.h5')

# evaluate the model
loss , accuracy =  model.evaluate(X, Y)
print('Accuracy :', accuracy*100)

from sklearn.metrics import precision_recall_curve
# import argmax
from numpy import argmax


#defining threshold as above
pred = model.predict(X)
pred = [1 if y>=0.5  else 0 for y in pred]
pred

X_features = X

import shap


# Using the SHAP library to explain the model's predictions

# Explainer ie shap.Explainer is created using the model and the data

# explainer = shap.Explainer(model, X_features)

# The SHAP values are calculated for the first 5 samples in the data
# shap_values = explainer(X_features)



# SAVING EXPLAINER
# joblib.dump(explainer, './Models/explainer.pkl')

# loading explainer

# explainer = joblib.load('./Models/explainer.pkl')

# explainer = shap.Explainer(model, X_features)

# # Find the feature that caused each anomaly
# if isinstance(shap_values, list):
#     # If shap_values is a list (e.g., for binary classification), use the second element
#     shap_df = pd.DataFrame(np.abs(shap_values[1]), columns=X_features.columns)
# else:
#     # If shap_values is already a 2D array
#     shap_df = pd.DataFrame(np.abs(shap_values.values), columns=X_features.columns[:7])


# # adding feature_cause column to the dataframe

# shap_df['feature_cause'] = shap_df.idxmax(axis=1)

# # saving the shap_df
# X_features['feature_cause'] = shap_df['feature_cause']


# df['feature_cause'] = X_features['feature_cause']

# df.to_csv('./DataSet/Test_1.csv', index=False)



X_features.head()


X_features.to_csv('./DataSet/Test.csv', index=False)

#Show the actual and predicted values
print('Original : {0}'.format(", ".join(str(x) for x in Y)))
print('Predicted : {0}'.format(", ".join(str(x) for x in pred)))


# Number of faulty ie 1's predicted by the model
faulty_predicted = pred.count(1)
print('Number of faulty predicted by the model :', faulty_predicted)

# Number of faulty ie 1's in the dataset
faulty_actual = Y.tolist().count(1)
print('Number of faulty in the dataset :', faulty_actual)


x = datetime.datetime.now()
print("Today's date: ")
print(x.strftime("%x"))

from datetime import date
from dateutil.relativedelta import relativedelta

six_months = date.today() + relativedelta(months=+6)
# print(six_months)
print("Scheduled Maintenance Date: ")
print(six_months.strftime("%x"))

from datetime import timedelta

# numpy array to pandas series
pred = np.array(pred)
n = (pred == 1).sum()
days_to_subtract = n/24 + 1
d = six_months - timedelta(days=days_to_subtract)
print("Predicted Maintenance Date: ")
print(d.strftime("%x"))

p = (24-n)/2400
print("Life of the machine is degraded by: ")
print(p)
