==> 6_deploy_model.ipynb <==
import os
import pandas as pd
from datetime import datetime, timedelta, timezone
from pymongo import MongoClient
from bson import ObjectId
from reportlab.lib import colors
from reportlab.lib.pagesizes import letter
from reportlab.platypus import SimpleDocTemplate, Paragraph, Spacer, Table, TableStyle
from reportlab.lib.styles import getSampleStyleSheet
from reportlab.lib.styles import getSampleStyleSheet
import pytz
import json
import matplotlib.pyplot as plt
import glob
from keras.models import Sequential, load_model
import numpy as np
from keras.layers import Dense
from sklearn.model_selection import train_test_split
from sklearn.preprocessing import LabelEncoder, MinMaxScaler
import keras as k
import datetime

import requests
import shutil

# Make new folder with name reading from model_data.json file

# Load model_data.json file

# reading from model_data.json file
# file {"mac": "08:D1:F9:A5:C8:84"}

with open('model_data.json') as json_file:
    data = json.load(json_file)
    mac = data['mac']


print(mac)

mac = mac.replace(":", "-")

# make folder with name of mac address

# delete folder if already exist

# mac/mac dir path

mac_dir = mac + "/" + mac

if os.path.exists(mac_dir):
    shutil.rmtree(mac_dir)

os.makedirs(mac_dir)


# copy model.h5 and scaler.pkl file from Model folder to mac folder

shutil.copy('Models/model.h5', mac_dir)
shutil.copy('Models/scaler.pkl', mac_dir)

# zip the folder

shutil.make_archive(mac, 'zip', mac)



# send the zip file to server

# API endpoint to upload model

url = "http://localhost:8000/update_model"

files = {'file': open(mac + '.zip', 'rb')}

response = requests.post(url, files=files)

print(response.json())

# remove zip file and folder

os.remove(mac + '.zip')

shutil.rmtree(mac)
